More code quality fixes

+++ src/object/floating_text.cpp
@@ -64,11 +64,11 @@ void
 FloatingText::draw(DrawingContext& context)
 {
   // make an alpha animation when disappearing
-  int alpha;
+  float alpha;
   if(timer.get_timeleft() < FADING_TIME)
-    alpha = int(timer.get_timeleft() * 255 / FADING_TIME);
+    alpha = timer.get_timeleft() * 255.0f / FADING_TIME;
   else
-    alpha = 255;
+    alpha = 255.0f;
 
   context.push_transform();
   context.set_alpha(alpha);
